r10885/sofa : push 3ba7512..863f934
* commit 863f9341f1f952a9d9fa3f864a75df2dae1a33f0
Move tests resources back to tests directories

They were moved into share/ in a previous commit, but this was not a
smart move (hehe). As it turns out, absolute paths are perfectly OK for
test resources, and defining a path as a macro in a CMakeLists.txt is
perfectly robust provided it is a string literal.

   34 files changed, 747 insertions(+), 742 deletions(-)




Former-commit-id: 11c55a1c608fbc5aeae19d8b867f318bd046a887