Backport scene files changes from branch unstable
Many corrections have been made to all the scene files in the "unstable"
branch (during the last coding sprint, and after).  This commit
backports them to the master branch.

By "backport", I mean a mere "cp". This is questionable with regards to
the commit log history, but we are also working into the master branch
to fix errors in the scene files, and merging is becoming painful.


Former-commit-id: cd4f9361180e82d8337612ba357f274a91f7c772